+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cam7: return movement or rotation to default by pressing both up+down together
 - Cam7: toggle hiding ui with `6`
 - Cam7: toggle disabling pod input with `7`
+- Cam7: move pod to camera by holding `BKSP(kb) or A(xinput)` when exiting free-cam
 - Cam7: sfx during camera motion
 - Core: notify on plugin reload
 - CollisionViewer: depth bias customization for correcting model-collision visual mismatch ([#5](https://github.com/everalert/annodue/pull/5))